+++ src/components/CartItemList/CartItemList.tsx
@@ -1,9 +1,9 @@
 import { useRecoilState } from 'recoil';
 import { useNavigate } from 'react-router-dom';
 import CheckButton from '../Button/CheckButton/CheckButton';
-import CartItemContainer from '../CartItem/CartItemContainer';
+import CartItemContainer from '../Container/CartItemContainer/CartItemContainer';
 import type { CartItem } from '../../types/CartItem.type';
-import { selectedCartItemListState } from '../../recoil/CartItem/atom/selectedCartItemListState';
+import { selectedCartItemListState } from '../../recoil/CartItem/atoms/selectedCartItemListState';
 import { removeCartItem, updateCartItemQuantity } from '../../apis';
 import * as S from './CartItemList.style';
