+++ src/pages/OrderConfirmPage/OrderConfirmPage.tsx
@@ -4,11 +4,11 @@ import Header from '../../components/Header/Header';
 import TitleContainer from '../../components/Container/TitleContainer/TitleContainer';
 import SubmitButton from '../../components/Button/SubmitButton/SubmitButton';
 import TotalPriceContainer from '../../components/Container/TotalPriceContainer/TotalPriceContainer';
-import CartItemContainer from '../../components/CartItem/CartItemContainer';
+import CartItemContainer from '../../components/Container/CartItemContainer/CartItemContainer';
 import DeliveryInfoContainer from '../../components/DeliveryInfoList/DeliveryInfoList';
 import ShowModalButton from '../../components/Button/ShowModalButton/ShowModalButton';
-import { selectedCartItemListState } from '../../recoil/CartItem/atom/selectedCartItemListState';
-import { selectedCartItemListTotalCountSelector } from '../../recoil/CartItem/selector/selectedCartItemListTotalCountSelector';
+import { selectedCartItemListState } from '../../recoil/CartItem/atoms/selectedCartItemListState';
+import { selectedCartItemListTotalCountSelector } from '../../recoil/CartItem/selectors/selectedCartItemListTotalCountSelector';
 import { PATHS } from '../../constants/PATHS';
 import * as S from './OrderConfirmPage.style';
 import { applyCouponModalState } from '../../recoil/ApplyCouponModal/atoms/applyCouponModalState';
