+++ tsconfig.json
@@ -24,6 +24,6 @@
     "types": ["jest"],
     "esModuleInterop": true
   },
-  "include": ["src", "src/recoil/CartItem/atom/selectedCartItemListState.ts"],
+  "include": ["src", "src/recoil/CartItem/atoms/selectedCartItemListState.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
